// src/features/groupsManagement/types/groups.types.ts
export interface ListResponse<T> {
    success: boolean;
    message: string;
    data: T[];
}